% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_brms.R
\name{adapted_brm}
\alias{adapted_brm}
\title{Adapted version of \code{brm()}}
\usage{
adapted_brm(
  formula,
  data,
  family = gaussian(),
  prior = NULL,
  autocor = NULL,
  data2 = NULL,
  cov_ranef = NULL,
  sample_prior = "no",
  sparse = NULL,
  knots = NULL,
  stanvars = NULL,
  stan_funs = NULL,
  fit = NA,
  save_ranef = TRUE,
  save_mevars = FALSE,
  save_all_pars = FALSE,
  inits = "random",
  chains = 4,
  iter = 2000,
  warmup = floor(iter/2),
  thin = 1,
  cores = getOption("mc.cores", 1L),
  control = NULL,
  algorithm = getOption("stan_algorithm", "sampling"),
  backend = getOption("stan_backend", "rstan"),
  future = getOption("future", FALSE),
  silent = TRUE,
  seed = NA,
  save_model = NULL,
  stan_model_args = list(),
  file = NULL,
  empty = FALSE,
  rename = TRUE,
  var_comp = NULL,
  ...
)
}
\description{
Similar to original \code{brms::brm()} function with some
    changes:
    \itemize{
        \item Takes argument \code{var_comp}
        \item Passes \code{var_comp} to function \code{\link{.make_stancode}}
        \item Passes \code{var_comp} to function \code{\link{.make_standata}}
    }
}
