% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit.R
\name{Bolker_exact_var_comp}
\alias{Bolker_exact_var_comp}
\title{Fit LMM with Exact Predefined Variance Components}
\usage{
Bolker_exact_var_comp(devfun, var_comp, lmod, mcout = quote(Bolker()))
}
\arguments{
\item{devfun}{Function which takes theta as input and returns the deviance.
Usually obtained from \code{\link{lme4:modular}[mkLmerDevfun]}}

\item{var_comp}{Ordered vector of variance components. Correct ordering can be
obtained with \code{\link{cowfit_var_comp}}.}

\item{lmod}{Output of \code{\link[lme4:lFormula]{lFormula}}}

\item{mcout}{Call which led to the output (\code{\link[base:match.call]{match.call}})}
}
\value{
Fitted model of class \code{\link[lme4:merMod-class]{merMod}}
}
\description{
Performs numerical optimization to find the
    scaling factor of \eqn{\theta} leading to a model with
    exact predefined variance components.
}
\examples{
lmod <- lFormula(formula = Reaction ~ Days + (Days | Subject),
                 data = sleepstudy)
devfun <- do.call(mkLmerDevfun, lmod)
(myvar <- cowfit_var_comp(formula = Reaction ~ Days + (Days | Subject),
                          data = sleepstudy))
myvar$vcov <- c(600, 30, 10, 600)
fit <- Bolker_exact_var_comp(devfun = devfun,
                             var_comp = myvar$vcov,
                             lmod = lmod)
as.data.frame(VarCorr(fit))
}
\seealso{
\code{\link{lme4:modular}[mkLmerDevfun]}, \code{\link{cowfit_var_comp}},
    \code{\link[lme4:lFormula]{lFormula}}, \code{\link[lme4:merMod-class]{merMod}}
}
\author{
Reto Zihlmann
}
