% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_brms.R
\name{cowfit_brm}
\alias{cowfit_brm}
\title{Fit bayesian GLMMs in genetic evaluation}
\usage{
cowfit_brm(
  formula,
  data,
  family = gaussian(),
  pedigree = list(),
  var_comp = NULL,
  prior = NULL,
  autocor = NULL,
  data2 = NULL,
  cov_ranef = NULL,
  sample_prior = "no",
  sparse = NULL,
  knots = NULL,
  stanvars = NULL,
  stan_funs = NULL,
  fit = NA,
  save_ranef = TRUE,
  save_mevars = FALSE,
  save_all_pars = FALSE,
  inits = "random",
  chains = 4,
  iter = 2000,
  warmup = floor(iter/2),
  thin = 1,
  cores = getOption("mc.cores", 1L),
  control = NULL,
  algorithm = getOption("stan_algorithm", "sampling"),
  backend = getOption("stan_backend", "rstan"),
  future = getOption("future", FALSE),
  silent = TRUE,
  seed = NA,
  save_model = NULL,
  stan_model_args = list(),
  file = NULL,
  empty = FALSE,
  rename = TRUE,
  ...
)
}
\arguments{
\item{formula}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{data}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{family}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{pedigree}{a named list of
\code{\link[pedigreemm:pedigree-class]{pedigree}} objects.
The names must correspond to the names of grouping factors
for random-effects terms in the formula argument.}

\item{var_comp}{Ordered vector of variance components. Correct ordering can be
obtained with \code{\link{cowfit_var_comp}}.}

\item{prior}{As in \code{\link[brms:brm]{brm}}.}

\item{autocor}{As in \code{\link[brms:brm]{brm}}.}

\item{data2}{As in \code{\link[brms:brm]{brm}}.}

\item{cov_ranef}{As in \code{\link[brms:brm]{brm}}.}

\item{sample_prior}{As in \code{\link[brms:brm]{brm}}.}

\item{sparse}{As in \code{\link[brms:brm]{brm}}.}

\item{knots}{As in \code{\link[brms:brm]{brm}}.}

\item{stanvars}{As in \code{\link[brms:brm]{brm}}.}

\item{stan_funs}{As in \code{\link[brms:brm]{brm}}.}

\item{fit}{As in \code{\link[brms:brm]{brm}}.}

\item{save_ranef}{As in \code{\link[brms:brm]{brm}}.}

\item{save_mevars}{As in \code{\link[brms:brm]{brm}}.}

\item{save_all_pars}{As in \code{\link[brms:brm]{brm}}.}

\item{chains}{As in \code{\link[brms:brm]{brm}}.}

\item{iter}{As in \code{\link[brms:brm]{brm}}.}

\item{warmup}{As in \code{\link[brms:brm]{brm}}.}

\item{thin}{As in \code{\link[brms:brm]{brm}}.}

\item{cores}{As in \code{\link[brms:brm]{brm}}.}

\item{control}{As in \code{\link[brms:brm]{brm}}.}

\item{algorithm}{As in \code{\link[brms:brm]{brm}}.}

\item{backend}{As in \code{\link[brms:brm]{brm}}.}

\item{future}{As in \code{\link[brms:brm]{brm}}.}

\item{silent}{As in \code{\link[brms:brm]{brm}}.}

\item{seed}{As in \code{\link[brms:brm]{brm}}.}

\item{save_model}{As in \code{\link[brms:brm]{brm}}.}

\item{stan_model_args}{As in \code{\link[brms:brm]{brm}}.}

\item{file}{As in \code{\link[brms:brm]{brm}}.}

\item{empty}{As in \code{\link[brms:brm]{brm}}.}

\item{rename}{As in \code{\link[brms:brm]{brm}}.}
}
\value{
Fitted object of class \code{\link[brms:brmsfit-class]{brmsfit}}
}
\description{
Fit Bayesian GLMMs with random effect correlated according to pedigree
    using Stan for full Bayesian inference.
}
\examples{
(myvar <- cowfit_var_comp(formula = y ~ (1|herd) + (protein|sire),
                          data = sim_milk))
myvar$vcov <- c(500, 400, 300, -200, 50)
fit <- cowfit_brm(y ~ protein + (1|herd) + (protein|sire),
                  data = sim_milk,
                  pedigree = list(sire = pedSires),
                  var_comp = c(500, 400, 300, -200, 50),
                  chains = 1)
}
\seealso{
\code{\link[brms:brm]{brm}}, \code{\link[cowfit_glmer]{cowfit_glmer}}
}
\author{
Reto Zihlmann
}
