% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_lme4.R
\name{cowfit_glmer}
\alias{cowfit_glmer}
\title{Fit GLMMs in genetic evaluation}
\usage{
cowfit_glmer(
  formula,
  data = NULL,
  family = gaussian,
  pedigree = list(),
  var_comp = NULL,
  cowfit_verbose = TRUE,
  control = glmerControl(),
  start = NULL,
  verbose = 0L,
  nAGQ = 1L,
  subset,
  weights,
  na.action,
  offset,
  contrasts = NULL,
  mustart,
  etastart,
  devFunOnly = FALSE
)
}
\arguments{
\item{formula}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{data}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{family}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{pedigree}{a named list of
\code{\link[pedigreemm:pedigree-class]{pedigree}} objects.
The names must correspond to the names of grouping factors
for random-effects terms in the formula argument.}

\item{var_comp}{Ordered vector of variance components. Correct ordering can be
obtained with \code{\link{cowfit_var_comp}}.}

\item{cowfit_verbose}{Logical. Verbose output is generated showing which module
of the function is currently active.}

\item{control}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{start}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{verbose}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{nAGQ}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{subset}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{weights}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{na.action}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{offset}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{contrasts}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{mustart}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{etastart}{As in \code{\link[lme4:glmer]{glmer}}.}

\item{devFunOnly}{As in \code{\link[lme4:glmer]{glmer}}.}
}
\value{
Fitted object of class \code{\link[glmercowfit-class]{glmercowfit}}
}
\description{
Fit GLMMs with random effect correlated according to pedigree.
}
\examples{
cowfit_var_comp(formula = y ~ lact + (lact|sire),
                data = sim_twin)
(myvar <- cowfit_var_comp(formula = y ~ lact + (lact|sire),
                          data = sim_twin))
myvar$vcov <- c(4, 3, -0.5, 1)
fit <- cowfit_glmer(formula = y ~ lact + (lact|sire),
                    data = sim_twin, family = "binomial",
                    pedigree = list(sire = pedSires),
                    var_comp = myvar$vcov)
as.data.frame(VarCorr(fit))
ranef(fit)
}
\seealso{
\code{\link[lme4:glmer]{glmer}}, \code{\link[pedigreemm:pedigreemm]{pedigreemm}}
}
\author{
Reto Zihlmann
}
