% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_brms.R
\name{data_gr_global}
\alias{data_gr_global}
\title{Adapted \code{data_gr_global}}
\usage{
data_gr_global(ranef, data2, var_comp = NULL)
}
\description{
Changes the orignial \code{data_gr_global} at two places.
    First, the function expects a cholesky factor and not a correlation matrix
    in \code{data2} i.e. no cholesky decomposition needs to be done.
    Second, the function adds the list of variance components to
    the output data.
}
\keyword{Internal}
