% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_brms.R
\name{.make_stancode}
\alias{.make_stancode}
\title{Adapted \code{.make_stancode()}}
\usage{
.make_stancode(
  bterms,
  data,
  prior,
  stanvars,
  parse = getOption("parse_stancode", FALSE),
  backend = getOption("stan_backend", "rstan"),
  silent = TRUE,
  save_model = NULL,
  var_comp = NULL,
  ...
)
}
\description{
Changes original \code{.make_stancode()} such that it takes argument
    \code{var_comp} and passes it to function \code{stan_re()}.
}
\keyword{Internal}
