% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_brms.R
\name{.make_standata}
\alias{.make_standata}
\title{Adapted \code{.make_standata()}}
\usage{
.make_standata(
  bterms,
  data,
  prior,
  stanvars,
  data2,
  check_response = TRUE,
  only_response = FALSE,
  internal = FALSE,
  basis = NULL,
  var_comp = NULL,
  ...
)
}
\description{
Changes original \code{.make_standata()} such that it takes argument
    \code{var_comp} and passes it to function \code{data_gr_global}.
}
\keyword{Internal}
