% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_brms.R
\name{.stan_re}
\alias{.stan_re}
\title{Adapted \code{.stan_re}}
\usage{
.stan_re(id, ranef, prior, var_comp = NULL)
}
\description{
Changes to original \code{.stan_re}
    \itemize{
        \item Takes argument var_comp
        \item Puts sd_i and L_i terms to data instead of parameters if \code{!is.null(var_comp)}
        \item Removes prior terms for sd_i and L_i if \code{!is.null(var_comp)}
    }
}
\keyword{Internal}
