% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_brms.R
\name{format_ranef}
\alias{format_ranef}
\title{Random effects in \code{lme4} format}
\usage{
format_ranef(object)
}
\arguments{
\item{object}{An object of a class of fitted models with random effects, including classes
\code{\link[lme4:merMod-class]{merMod}}, \code{\link{cowfit}} and
\code{\link[brms:brmsfit-class]{brmsfit}}.}
}
\value{
List of data frames with the random effects of each term
}
\description{
Returns the random effect of a fitted model in the \code{lme4}
    format.
}
\examples{
    fit <- cowfit_lmer(y ~ (1|sire), data = sim_milk,
                       pedigree = list(sire = pedSires))
    format_ranef(fit)
}
\seealso{
\code{\link[lme4:merMod-class]{merMod}}, \code{\link{cowfit}}, \code{\link[brms:brmsfit-class]{brmsfit}}
}
\author{
Reto Zihlmann
}
