% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_lme4.R
\name{get_TAt}
\alias{get_TAt}
\title{Get transformation factor}
\usage{
get_TAt(lmod, pedigree, include_TAt_list = FALSE)
}
\arguments{
\item{lmod}{Output of \code{\link[lme4:lFormula]{lFormula}}}

\item{pedigree}{a named list of
\code{\link[pedigreemm:pedigree-class]{pedigree}} objects.
The names must correspond to the names of grouping factors
for random-effects terms in the formula argument.}

\item{include_TAt_list}{Logical. Should also the term-wise separated TAt (TAt_list) be returend?}
}
\value{
Transformation factor
}
\description{
Transformation factor is used to transform the model
    matrix \eqn{Z^T} in order to make the random effects independent
    between animals.
}
\examples{
lmod <- lFormula(y ~ (1|herd) + (protein|sire),
                 data = sim_milk)
get_TAt(lmod = lmod, pedigree = list(sire = pedSires))
}
\seealso{
\code{\link[lme4:lFormula]{lFormula}}, \code{\link[pedigreemm:pedigree-class]{pedigree}}
}
\author{
Reto Zihlmann
}
