% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse Logit function}
\usage{
inv_logit(eta)
}
\arguments{
\item{eta}{logarithm of the odds [-inf, +inf]}
}
\value{
probability [0, 1]
}
\description{
Calculates Probability from logarithm of the odds
}
\examples{
inv_logit(12)
}
