% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_lme4.R
\docType{class}
\name{lme4cowfit-class}
\alias{lme4cowfit-class}
\title{Fitted Pedigree-based mixed effects model}
\description{
A mixed effects model fit by \code{\link{cowfit_lmer}} or \code{\link{cowfit_glmer}}.
    This class extends class \code{\link[lme4:merMod-class]{merMod}} and includes one
    additional slot, \code{TAt}.
}
\section{Slots}{

\describe{
\item{\code{TAt}}{sparse matrix of class \code{\link[Matrix:dtCMatrix-class]{dtCMatrix}},
which was used to transform the
model matrix \eqn{Z^T} and can be used to backtransform the random effects.}
}}

\examples{
showClass("lme4cowfit")
}
\seealso{
\code{\link[cowfit:glmercowfit-class]{glmercowfit}}, \code{\link[cowfit:lmercowfit-class]{lmercowfit}}
}
