% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_lme4.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{probability [0, 1]}
}
\value{
Logarithm of the odds  [-inf, +inf]
}
\description{
Calculates Logarithm of the odds
}
\examples{
logit(0.75)
}
