% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit.R
\name{sim_glmer}
\alias{sim_glmer}
\title{Simulate GLMMs in genetic evaluation}
\usage{
sim_glmer(
  formula = y ~ (1 | herd) + (lact | id),
  data = milk,
  family = "binomial",
  pedigree = list(),
  var_comp,
  beta,
  return_ranef = TRUE
)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame containing all predictors in \code{formula}}

\item{family}{A GLM family, see \code{\link[stats:glm]{glm}}
and \code{\link[stats:family]{family}}.}

\item{pedigree}{a named list of
\code{\link[pedigreemm:pedigree-class]{pedigree}} objects.
The names must correspond to the names of grouping factors
for random-effects terms in the formula argument.}

\item{var_comp}{Ordered vector of variance components. Correct ordering can be
obtained with \code{\link{cowfit_var_comp}}.}

\item{beta}{Ordered vector of fixed effects}

\item{retrun_ranef}{Logical. Should true random effects returned together with data}
}
\value{
List containing simulated data (and optionally random effects)
}
\description{
Simulate GLMM with genetically correlated random effects
}
\examples{
set.seed(1)
sim_glmer(formula = y ~ lact + (lact|sire), data = sim_twin,
          family = "binomial", pedigree = list(sire = pedSires),
          var_comp = c(4, 3, -0.5, 1), beta = c(-3, 0.5),
          return_ranef = FALSE)
}
\seealso{
\code{\link{cowfit_glmer}}
}
\author{
Reto Zihlmann
}
