% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_lme4.R
\name{sim_lmer}
\alias{sim_lmer}
\title{Simulate LMMs in genetic evaluation}
\usage{
sim_lmer(
  formula = y ~ (1 | herd) + (lact | id),
  data = milk,
  pedigree = list(),
  var_comp,
  beta,
  return_ranef = TRUE
)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame containing all predictors in \code{formula}}

\item{pedigree}{a named list of
\code{\link[pedigreemm:pedigree-class]{pedigree}} objects.
The names must correspond to the names of grouping factors
for random-effects terms in the formula argument.}

\item{var_comp}{Ordered vector of variance components. Correct ordering can be
obtained with \code{\link{cowfit_var_comp}}.}

\item{beta}{Ordered vector of fixed effects}

\item{retrun_ranef}{Logical. Should true random effects returned together with data}
}
\value{
List containing simulated data (and optionally random effects)
}
\description{
Simulate LMM with genetically correlated random effects
}
\examples{
set.seed(1)
sim_lmer(formula = y ~ protein + (1|herd) + (protein|sire),
         data = sim_milk,
         pedigree = list(sire = pedSires),
         var_comp = c(500, 400, 300, -200, 50),
         beta = c(6000, 200), return_ranef = FALSE)
}
\seealso{
\code{\link{cowfit_lmer}}
}
\author{
Reto Zihlmann
}
