% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sim_milk}
\alias{sim_milk}
\title{Simulated effect of protein on milk yield}
\format{
An object of class \code{data.frame}
}
\usage{
data(sim_milk)
}
\description{
Simulated milk yield in population of pedigree \code{\link[pedigreemm:pedSires]{pedSires}}.
Example for LMMs. Simulation based on model \code{y ~ protein + (1|herd) + (protein|sire)} with
\code{beta = c(6000, 200)} and \code{var_comp = 500  400  300 -200   50}
}
\examples{
data(sim_milk)
}
\keyword{datasets}
