% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sim_twin}
\alias{sim_twin}
\title{Simulated effect of lactation on multiple birth}
\format{
An object of class \code{data.frame}
}
\usage{
data(sim_twin)
}
\description{
Simulated birth in population of pedigree \code{\link[pedigreemm:pedSires]{pedSires}}.
Example for GLMMs. Simulation based on model \code{y ~ lact + (lact|sire)} with
\code{beta = c(-3, 0.5)} and \code{var_comp = c(4, 3, -0.5, 1)}. \code{y = 0} stands for
single birth and \code{y = 1} for multiple birth.
}
\examples{
data(sim_twin)
}
\keyword{datasets}
