% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_adapted.R
\name{stan_adapt}
\alias{stan_adapt}
\title{stan_glmer() which transformes Z matrix with Lt}
\usage{
stan_adapt(
  formula,
  data = NULL,
  family = gaussian,
  subset,
  weights,
  na.action = getOption("na.action", "na.omit"),
  offset,
  contrasts = NULL,
  ...,
  prior = normal(),
  prior_intercept = normal(),
  prior_aux = exponential(),
  prior_covariance = decov(),
  prior_PD = FALSE,
  algorithm = c("sampling", "meanfield", "fullrank"),
  adapt_delta = NULL,
  QR = FALSE,
  sparse = FALSE
)
}
\description{
stan_glmer() which transformes Z matrix with Lt
}
