% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_brms.R
\name{stan_re}
\alias{stan_re}
\title{Adapted \code{stan_re}}
\usage{
stan_re(ranef, prior, var_comp = NULL, ...)
}
\description{
Changes original \code{stan_re()} such that it takes argument
    \code{var_comp} and passes it to function \code{.stan_re()}.
}
\keyword{Internal}
