% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_brms.R
\name{test_cowfit_brm}
\alias{test_cowfit_brm}
\title{Test fitting function}
\usage{
test_cowfit_brm(
  formula = ~lact + (1 | herd) + (lact | sire),
  data = milk,
  pedigree = list(sire = pedCowsR),
  family = "gaussian",
  var_comp = c(4, 6, 2, 0.5, 1),
  beta = c(100, 2),
  given_var_comp = TRUE,
  return_all = TRUE,
  ...
)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame containing all predictors in \code{formula}}

\item{pedigree}{a named list of
\code{\link[pedigreemm:pedigree-class]{pedigree}} objects.
The names must correspond to the names of grouping factors
for random-effects terms in the formula argument.}

\item{family}{A GLM family, see \code{\link[stats:glm]{glm}}
and \code{\link[stats:family]{family}}.}

\item{var_comp}{Ordered vector of variance components. Correct ordering can be
obtained with \code{\link{cowfit_var_comp}}.}

\item{beta}{Ordered vector of fixed effects}

\item{given_var_comp}{Logical. Should true variance components be passed to
fitting function \code{\link{cowfit_glmer}}?}

\item{return_all}{Logical. If \code{TRUE} not only fitted model but also
\code{\link{system.time}} and correlation between true and fitted random effects.}
}
\value{
list containing fitted model (and optionally \code{\link{system.time}} and
    correlation between true and fitted random effects.)
}
\description{
Test \code{\link{cowfit_brm}} for speed and accuracy of predicted
    random effects.
}
\examples{
cowfit_var_comp(formula = y ~ lact + (lact|sire), data = sim_twin)
out <- test_cowfit_brm(formula = y ~ lact + (lact|sire), data = sim_twin,
                       family = "binomial", pedigree = list(sire = pedSires),
                       var_comp = c(4, 3, -0.5, 1), beta = c(-3, 0.5),
                       given_var_comp = TRUE)
}
\seealso{
\code{\link{cowfit_glmer}}, \code{\link{sim_glmer}}
}
\author{
Reto Zihlmann
}
