% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_lme4.R
\name{test_cowfit_lmer}
\alias{test_cowfit_lmer}
\title{Test fitting function}
\usage{
test_cowfit_lmer(
  formula = ~lact + (1 | herd) + (lact | sire),
  data = milk,
  pedigree = list(sire = pedCowsR),
  var_comp = c(4, 6, 2, 0.5, 1),
  beta = c(100, 2),
  given_var_comp = TRUE,
  exact_var_comp = FALSE,
  return_all = TRUE,
  ...
)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame containing all predictors in \code{formula}}

\item{pedigree}{a named list of
\code{\link[pedigreemm:pedigree-class]{pedigree}} objects.
The names must correspond to the names of grouping factors
for random-effects terms in the formula argument.}

\item{var_comp}{Ordered vector of variance components. Correct ordering can be
obtained with \code{\link{cowfit_var_comp}}.}

\item{beta}{Ordered vector of fixed effects}

\item{given_var_comp}{Logical. Should true variance components be passed to
fitting function \code{\link{cowfit_lmer}}?}

\item{exact_var_comp}{Logical. Should numerical optimization be used to find exact theta}

\item{return_all}{Logical. If \code{TRUE} not only fitted model but also
\code{\link{system.time}} and correlation between true and fitted random effects.}
}
\value{
list containing fitted model (and optionally \code{\link{system.time}} and
    correlation between true and fitted random effects.)
}
\description{
Test \code{\link{cowfit_lmer}} for speed and accuracy of predicted
    random effects.
}
\examples{
cowfit_var_comp(formula = ~ calvingYear + (1|herd) + (mastitis|sire),
                data = mastitis)
out <- test_cowfit_lmer(formula = ~ calvingYear + (1|herd) + (mastitis|sire),
                        data = mastitis, pedigree = list(sire = pedSires),
                        var_comp = c(4, 5, 2, 0.5, 0.5),
                        beta = c(100, 2, 1, 3, 4, 3),
                        given_var_comp = TRUE, exact_var_comp = TRUE)
as.data.frame(VarCorr(out$fit))
}
\seealso{
\code{\link{cowfit_lmer}}, \code{\link{sim_lmer}}
}
\author{
Reto Zihlmann
}
