% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_brms.R
\name{var_comp_to_stan_format}
\alias{var_comp_to_stan_format}
\title{Transform variance component vector to list for rstan}
\usage{
var_comp_to_stan_format(var_comp, formula, data)
}
\arguments{
\item{var_comp}{Ordered vector of variance components. Correct ordering can be
obtained with \code{\link{cowfit_var_comp}}.}

\item{formula}{Model formula. As in \code{\link[lme4:lmer]{lmer}}.}

\item{data}{As in \code{\link[lme4:lmer]{lmer}}.}
}
\value{
List containing variance components used as input for \code{\link[rstan:sampling]{rstan::sampling}}
}
\description{
Variance components need to be passed to rstan in the form of
    a list containing the standard deviations as \code{sd_i} terms and
    cholesky factors of correlation matrices as \code{L_i} terms.
}
\examples{
    var_comp <- var_comp_to_stan_format(var_comp = c(500, 400, 300, -200, 50),
                                        formula = y ~ protein + (1|herd) + (protein|sire),
                                        data = sim_milk)
}
\seealso{
\code{\link[rstan:sampling]{rstan::sampling}}
}
\author{
Reto Zihlmann
}
