% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_lme4.R
\name{var_to_theta}
\alias{var_to_theta}
\title{Transform var_comp vector to theta}
\usage{
var_to_theta(var_comp, cnms)
}
\arguments{
\item{var_comp}{Ordered vector of variance components. Correct ordering can be
obtained with \code{\link{cowfit_var_comp}}.}

\item{cnms}{List with column names of raw random effects model matrix \eqn{X_i}.
Usually obtained from output of \code{\link[lme4:lFormula]{lFormula}}.}
}
\value{
\code{theta}
}
\description{
Theta is the vector which is used for numerical optimization in \code{lme4}.
    Theta contains the non-zero elements of the left Cholesky factor of Sigma.
    Sigma is the variance covarince matrix of the random effects.
}
\examples{
lmod <- lFormula(y ~ (1|herd) + (protein|sire), data = sim_milk)
var_to_theta(var_comp = c(500, 400, 300, -200, 50),
             cnms = lmod$reTrms$cnms)
}
\seealso{
\code{\link[lme4:lFormula]{lFormula}}
}
\author{
Reto Zihlmann
}
