% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_lme4.R
\docType{class}
\name{glmercowfit-class}
\alias{glmercowfit-class}
\title{Fitted Pedigree-based GLMM}
\description{
A mixed effects model fit by \code{\link{cowfit_glmer}}.
    The class inherits from class \code{\link[cowfit:lme4cowfit-class]{lme4cowfit}}.
}
\examples{
showClass("lmercowfit")
}
\seealso{
\code{\link[cowfit:lme4cowfit-class]{lme4cowfit}}, \code{\link[cowfit:lmercowfit-class]{lmercowfit}}
}
