% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowfit_lme4.R
\docType{class}
\name{lmercowfit-class}
\alias{lmercowfit-class}
\title{Fitted Pedigree-based LMM}
\description{
A mixed effects model fit by \code{\link{cowfit_lmer}}.
    The class inherits from class \code{\link[cowfit:lme4cowfit-class]{lme4cowfit}}.
}
\examples{
showClass("lmercowfit")
}
\seealso{
\code{\link[cowfit:lme4cowfit-class]{lme4cowfit}}, \code{\link[cowfit:glmercowfit-class]{glmercowfit}}
}
